\name{computeIncidence}
\alias{computeIncidence}
\title{Compute the average incidence rate}
\usage{
  computeIncidence(sim)
}
\arguments{
  \item{sim}{see \code{\link{doSim}}}
}
\description{
  The incidence rate is the average number of infections
  present in the hosiptal. I compute it as (total infected
  patient days)/(total hospital patient days)
}

