% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{event.levels}
\alias{event.levels}
\title{event levels for consistency accross functions.}
\format{\preformatted{ chr [1:5] "admit" "negative" "positive" "discharge" "infection"
}}
\usage{
event.levels
}
\description{
event levels for consistency accross functions.
}
\keyword{datasets}

