% Generated by roxygen2 (4.0.0): do not edit by hand
\name{formatTime}
\alias{formatTime}
\title{formats proc_time as a "hh:mm:ss" time string}
\usage{
formatTime(x)
}
\arguments{
\item{x}{number of seconds.}
}
\description{
formats proc_time as a "hh:mm:ss" time string
}
\seealso{
print.proc_time, proc.time
}

