\name{mcmc_chain_plot}
\alias{mcmc_chain_plot}
\title{Create a plot of the chains of parameters.}
\usage{
  mcmc_chain_plot(run,
    vars = c("transmission", "importation", "false.neg"))
}
\arguments{
  \item{run}{output from run method of a
  \code{\link{cont.inf.model}} object.}

  \item{vars}{variables to plot.}
}
\description{
  Create a plot of the chains of parameters.
}
\keyword{plot}

