% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{transmission}
\alias{transmission}
\alias{transmission-package}
\title{Infectious Disease Transmission Modeling on Individuals}
\description{
The transmission package provides methods for simulating data from a transmission
 model in continuous time.
}
\details{


The model is fit through a continuous time MCMC estimation.
}
\author{
Andrew Redd \email{Andrew.Redd@hsc.utah.edu},
 Alun Thomas \email{alun.thomas@utah.edu},
 Karim Khadar \email{Karim.Khader@hsc.utah.edu}

 Maintainer: Andrew Redd \email{Andrew.Redd@hsc.utah.edu}
}
\seealso{
see \code{\link{doSim}} for simulating data.
}
\keyword{package}

