% Generated by roxygen2 (4.0.0): do not edit by hand
\name{Vswitch}
\alias{Vswitch}
\title{Vectorized switch}
\usage{
Vswitch(EXPR, ...)
}
\arguments{
\item{EXPR}{the expression to vectorize over.}

\item{...}{passed onto switch for cases evaluation.}
}
\description{
Vectorized switch
}
\seealso{
\code{\link{switch}}
}

