\name{cont.inf.model}
\alias{cont.inf.model}
\alias{cont.inf.model-class}
\title{Continuous Time Infectious Disease Modeling on Individuals}
\arguments{
  \item{df}{an event formatted data.frame.}

  \item{patients}{an data frame of patient information with
  admission, discharge and optionally infection times}

  \item{tests}{an data frame of test information with test
  time, patient id, and result.}
}
\description{
  Continuous Time Infectious Disease Modeling on
  Individuals
}
\seealso{
  \code{\link{doSim}} for simulating data.

  Other mcmc: \code{\link{tmMCMC}}
}

