\docType{data}
\name{event.levels}
\alias{event.levels}
\title{event levels for consistency accross functions.}
\format{chr [1:5] "admit" "negative" "positive" "discharge" "infection"}
\usage{
  event.levels
}
\description{
  event levels for consistency accross functions.
}
\keyword{datasets}

