% Generated by roxygen2 (4.0.0): do not edit by hand
\name{print.proc_time}
\alias{print.proc_time}
\title{Prints out time in an intelligent manner.}
\usage{
\method{print}{proc_time}(x, ...)
}
\arguments{
\item{x}{object of class proc_time}

\item{...}{ignored}
}
\description{
Prints out time in an intelligent manner.
}
\seealso{
system.time proc.time
}

