\docType{package}
\name{transmission}
\alias{transmission}
\alias{transmission-package}
\title{Infectious Disease Transmission Modeling on Individuals}
\description{
  Transmission Model Simulation and Fitting in Continuous
  Time for Small Populations
}
\details{
  The transmission package provides methods for simulating
  data from a transmission model in continuous time.
}
\author{
  Andrew Redd \email{Andrew.Redd@hsc.utah.edu}, Alun Thomas
  \email{alun.thomas@utah.edu}, Karim Khadar
  \email{Karim.Khader@hsc.utah.edu}

  Maintainer: Andrew Redd \email{Andrew.Redd@hsc.utah.edu}
}
\seealso{
  see \code{\link{doSim}} for simulating data.
}
\keyword{package}

